Revert "New teams peoples actions"
