--- src/datos.py
@@ -1,5 +1,0 @@
-import utils
-
-teams = utils.load_from_json('src/teams/teams.json')
-tasks = utils.load_from_json('src/tasks/tasks.json')
-people = utils.load_from_json('src/people/people.json')
