--- src/people/people.json
@@ -1,1 +1,0 @@
-[{"name": "Arsenii", "surname": "Abramos", "age": 19, "specialization": "Junior programador"}, {"name": "Daniil", "surname": "Igoshev", "age": 19, "specialization": "Team lid"}, {"name": "Hui ", "surname": "Dolbaebov", "age": 18, "specialization": "Middle programador"}]
