--- src/tasks/tasks.json
@@ -1,1 +1,0 @@
-[]
