--- src/teams/teams.json
@@ -1,1 +1,0 @@
-[{"name": "123", "persons": [{"name": "Arsenii", "surname": "Abramos", "age": 19, "specialization": "Junior programador"}, {"name": "Daniil", "surname": "Igoshev", "age": 19, "specialization": "Team lid"}]}]
